requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
